chatbot-mcp/mcp-server/requirements.txt
# This file is kept for backward compatibility
# For new projects, use pyproject.toml and uv instead
# To install: uv sync (recommended) or pip install -r requirements.txt

mcp>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0